Pave the way for custom formatters
Something has been bugging me about the fact that ObjectIdentifier is
so opinionated about its output given the name is fairly generic. So
I've gone ahead and setup a basis for different Formatters, the default
of which will be the String-style formatter/output. A future iteration
will bring a configuration option for what is the default formatter.

The ability to define custom formatters may be somewhat academic. It
does loosen the coupling the ObjectIdentifier::Identifier has, though,
by introducing dependency injection. This is the spirit of OO and, since
this is a gem, I think it should seek to satisfy that spirit.

Also, reduce some conditional logic between multi-object and
single-object to-be-identified entities.

Finally, fix up rubocop offenses and style decisions to my latest style
guide after these code updates.