Split StringFormatter implementation into separate concerns
Behavior for single vs multiple objects is really composed of a special
wrapper for multiple objects + single object formatting. We can pass
around state fewer times by decomposing these differences into separate
objects. This reads cleaner and satisfies my pure OO itch.

Note: There is a cost for this change in terms of speed, although it
isn't large:

Before:
  58.478k (± 0.8%) i/s -    295.776k in   5.058178s
After:
  47.115k (± 0.3%) i/s -    238.935k in   5.071329s

Could build a speed-optimized StringFormatter if this is a problem...
or something else. Will thin on it, but for now the code is easier to
follow and more open to change, I believe.