Reduce attributes list passing
OO says that if you're passing around a number of attributes then
really that's begging for an object to be created to encapsulate that
concept. So, this effort teases this concept out by creating a
ObjectIdentifier::Identifier::Parameters object to represent the
repeatedly passed around params: `attributes`, `limit`, and `klass`.

This was an interesting exercise, but was a much bigger change than
I had expected up front. Primarily because almost all of the tests
needed to be adapted as well since I'm trying to unit test
StringFormatter in isolation (without going through the main
ObjectIdentifier::Identifier.call method, which would otherwise abstract
most of these interface-only changes).

Changing interfaces is messy, though... because changing a concretion
changes all dependencies. I'm not sure there was a better way to change
a little at a time using red/green testing to get this done... but I
should like to explore that in the future.