Update RuboCop config to source from pdobb-style gem
The new pdobb-style gem is an attempt to commonize my style guide across
all of my gems and/or future projects.

After updating to the new common base I had to fix a few issues in the
existing code as well, coming from inclusion of the rubocop-performance
gem that's now being included.

Also, rename the `scripts` folder to `script` to follow along with what
is expected by the new style guide (gem).