Refactor option `klass` -> `class`
I can never remember what it's supposed to be... I haven't used `klass`
as my convention in a long time now.

For backwards compatibility, we still allow the `klass` option, but we
emit a DEPRECATION WARNING.

Update the README to use real objects and to show outputs as Strings
(which they are, even if that's a bit messier).